+++ docs/markdown/ReleaseNotes.md
@@ -2,10 +2,12 @@
 
 ## Version 2.10.0
 
-* additional `NdArray` access operator overloads and `at()` overloads
-* additional `put()` overloads to `NdArray`
-* fixed **Issue #170**: slice and put with various integer index types
+* additional `NdArray` access operator overloads and `at` overloads
+* additional `put` overloads to `NdArray`
+* added `dimSize` method to `NdArray`
 * added `timeit` function
+* added `insert` function for **Issue #170** <https://numpy.org/doc/stable/reference/generated/numpy.insert.html>
+* fixed **Issue #177**: slice and put with various integer index types
 * various performance improvements
 
 ## Version 2.9.0